+++ src/main/java/com/nextech/erp/model/Qualitycheckrawmaterial.java
@@ -27,7 +27,7 @@ public class Qualitycheckrawmaterial implements Serializable {
 	private long id;
 
 	@Column(name="created_by")
-	private int createdBy;
+	private long createdBy;
 
 	@Column(name="created_date")
 	private Timestamp createdDate;
@@ -43,7 +43,7 @@ public class Qualitycheckrawmaterial implements Serializable {
 	private String remark;
 
 	@Column(name="updated_by")
-	private int updatedBy;
+	private long updatedBy;
 
 	@Column(name="updated_date")
 	private Timestamp updatedDate;
@@ -83,11 +83,11 @@ public class Qualitycheckrawmaterial implements Serializable {
 		this.id = id;
 	}
 
-	public int getCreatedBy() {
+	public long getCreatedBy() {
 		return this.createdBy;
 	}
 
-	public void setCreatedBy(int createdBy) {
+	public void setCreatedBy(long createdBy) {
 		this.createdBy = createdBy;
 	}
 
@@ -139,11 +139,11 @@ public class Qualitycheckrawmaterial implements Serializable {
 		this.remark = remark;
 	}
 
-	public int getUpdatedBy() {
+	public long getUpdatedBy() {
 		return this.updatedBy;
 	}
 
-	public void setUpdatedBy(int updatedBy) {
+	public void setUpdatedBy(long updatedBy) {
 		this.updatedBy = updatedBy;
 	}
