+++ src/main/java/com/nextech/erp/model/Rawmaterialorderinvoiceassociation.java
@@ -21,7 +21,7 @@ public class Rawmaterialorderinvoiceassociation implements Serializable {
 	private long id;
 
 	@Column(name="created_by")
-	private int createdBy;
+	private long createdBy;
 
 	@Column(name="created_date")
 	private Timestamp createdDate;
@@ -29,7 +29,7 @@ public class Rawmaterialorderinvoiceassociation implements Serializable {
 	private boolean isactive;
 
 	@Column(name="updated_by")
-	private int updatedBy;
+	private long updatedBy;
 
 	@Column(name="updated_date")
 	private Timestamp updatedDate;
@@ -58,11 +58,11 @@ public class Rawmaterialorderinvoiceassociation implements Serializable {
 		this.id = id;
 	}
 
-	public int getCreatedBy() {
+	public long getCreatedBy() {
 		return this.createdBy;
 	}
 
-	public void setCreatedBy(int createdBy) {
+	public void setCreatedBy(long createdBy) {
 		this.createdBy = createdBy;
 	}
 
@@ -82,11 +82,11 @@ public class Rawmaterialorderinvoiceassociation implements Serializable {
 		this.isactive = isactive;
 	}
 
-	public int getUpdatedBy() {
+	public long getUpdatedBy() {
 		return this.updatedBy;
 	}
 
-	public void setUpdatedBy(int updatedBy) {
+	public void setUpdatedBy(long updatedBy) {
 		this.updatedBy = updatedBy;
 	}
